% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc.R
\docType{data}
\name{crc}
\alias{crc}
\title{Data for Metastatic Colorectal Cancer clinical trials}
\format{
\subsection{\code{crc}}{

A data frame with 717 rows and 12 columns:
\describe{
\item{nct_id}{the identifier of the clinical trials.}
\item{tudy_type}{interventional or observations.}
\item{brief_title}{brief description of the clinical trial.}
\item{phase}{study phase.}
}
}
}
\usage{
crc
}
\description{
Data for Metastatic Colorectal Cancer clinical trials
}
\keyword{datasets}
