% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc_sub.R
\docType{data}
\name{crc_sub}
\alias{crc_sub}
\title{Data for Metastatic Colorectal Cancer clinical trials with specific interventions}
\format{
\subsection{\code{crc_sub}}{

A data frame with 1697 rows and 16 columns:
\describe{
\item{nct_id}{the identifier of the clinical trials.}
\item{intervention_type}{specific types of interventions (e.g., Drug, Procedure, Biological, etc.).}
}
}
}
\usage{
crc_sub
}
\description{
Data for Metastatic Colorectal Cancer clinical trials with specific interventions
}
\keyword{datasets}
