% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{data_query_search}
\alias{data_query_search}
\title{Function that filters the studies data based on multiple queries,
including keyword matching and date range filtering}
\usage{
data_query_search(studies, kws, dates)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{kws}{A string. The keyword string entered by the users}

\item{dates}{A list of two date values (start and end date)}
}
\value{
A tibble with filtered entries
}
\description{
Function that filters the studies data based on multiple queries,
including keyword matching and date range filtering
}
