% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_analysis.R
\name{duration_regression_analysis}
\alias{duration_regression_analysis}
\title{Regression Analysis of Study Duration}
\usage{
duration_regression_analysis(dates = c("2000-01-01", "2100-01-01"))
}
\arguments{
\item{dates}{A list of two date values (start and end date, respectively) to define the period for analysis.}
}
\value{
The function returns the results from the regression analysis, diagnostic plots to assess the model, and a plot comparing
actual vs predicted values of study duration.
}
\description{
This function performs regression analysis on the duration of studies, considering factors such as enrollment, source class, and phase. The analysis includes the removal of outliers based on Cook's distance, and is conducted within a specified date range.
}
