% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_duration_analysis.R
\name{enrollment_duration_analysis}
\alias{enrollment_duration_analysis}
\title{Statistical Analysis of Enrollment Across Trials with Duration of Study}
\usage{
enrollment_duration_analysis(dates = c("2000-01-01", "2100-01-01"))
}
\arguments{
\item{dates}{A list of two date values (lower and upper, respectively) defining the period for analysis.}
}
\value{
Returns the results from the ANOVA analysis and a scatter plot showing the relationship between study duration and enrollment counts.
}
\description{
This function performs a statistical analysis to explore the relationship between enrollment counts and study duration in colorectal cancer trials within a specified date range. The analysis includes ANOVA and the creation of a scatter plot for visualizing the relationship.
}
