% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_source_analysis.R
\name{enrollment_source_analysis}
\alias{enrollment_source_analysis}
\title{Statistical analysis of enrollment across trials with different source classes}
\usage{
enrollment_source_analysis(dates = c("2000-01-01", "2100-01-01"))
}
\arguments{
\item{dates}{A list of two date values (lower and upper, respectively)}
}
\value{
Results from the ANOVA and Kruskal-Wallis test, enrollment distribution
histogram, and barp lot showing mean enrollment counts across different source classes
}
\description{
Given a specific date range, this functions explores the
differences in enrollment counts across different colorectal cancer trials
with different source classes, by performing ANOVA and Kruskal-Wallis and
generating associated visualizations.
}
