% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crc_sankey.R
\name{plot_crc_sankey}
\alias{plot_crc_sankey}
\title{Sankey plot of trial study types, intervention types, and phases}
\usage{
plot_crc_sankey(dates = c("2000-01-01", "2100-01-01"))
}
\arguments{
\item{dates}{A list of two date values (lower and upper, respectively)}
}
\value{
A Sankey (or Alluvial) plot showing the clinical trials distribution
}
\description{
Filters mCRC clinical trials within a given range of trial start date
and visualizes a Sankey diagram showing the distribution of study types stratified
by intervention types and phases
}
