% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intervention_prop.R
\name{plot_intervention_proportion_time}
\alias{plot_intervention_proportion_time}
\title{Proportions of differnt intervention types over time}
\usage{
plot_intervention_proportion_time(dates = c("2000-01-01", "2100-01-01"))
}
\arguments{
\item{dates}{A list of two date values (lower and upper, respectively)}
}
\value{
A density plot showing the proportion of intervention types vs. time
}
\description{
Filters mCRC clinical trials within a given range of trial start date
and plots a density plot showing the proportion of each intervention type over time
}
