% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_query_time_trend.R
\name{plot_query_time_trend}
\alias{plot_query_time_trend}
\title{Function that query the clinical trial table and plots the number of trials
matching the given keywords over time.}
\usage{
plot_query_time_trend(kwds, column, ignore.case = TRUE, match_all = FALSE)
}
\arguments{
\item{kwds}{A string / list of strings. The query keyword(s)}

\item{column}{A quosure (a string enclosed in \code{quo()}). The column to be filtered upon.}

\item{match_all}{A Boolean, default FALSE. Specify whether the query
requires the column value to match all the kwds or just one of them}

\item{ignore_case}{A Boolean, default TRUE. Specify whether the query is case-insensitive}
}
\value{
a line plot showing the count of queried clinical trials over time
}
\description{
Function that query the clinical trial table and plots the number of trials
matching the given keywords over time.
}
