% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_query_time_trend.R
\name{query_kwds}
\alias{query_kwds}
\title{Function that queries the \code{crc_sub} table using a keyword/a list of keyword strings}
\usage{
query_kwds(kwds, column, ignore.case = TRUE, match_all = FALSE)
}
\arguments{
\item{kwds}{A string / list of strings. The query keyword(s)}

\item{column}{A quosure (a string enclosed in \code{quo()}). The column to be filtered upon.}

\item{ignore.case}{A Boolean, default TRUE. Specify whether the query is case-insensitive}

\item{match_all}{A Boolean, default FALSE. Specify whether the query
requires the column value to match all the kwds or just one of them}
}
\value{
A table with filtered entries
}
\description{
Function that queries the \code{crc_sub} table using a keyword/a list of keyword strings
}
