% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{status_classification}
\alias{status_classification}
\title{Overall Status classification task}
\usage{
status_classification(dates = c("2000-01-01", "2100-01-01"))
}
\arguments{
\item{dates}{A list of two date values (lower and upper, respectively)}
}
\value{
Results from the logistic regression model, including a confusion matrix,
accuracy levels, and a graph visualizing the confusion matrix values.
}
\description{
Given a specific date range, this functions classifies overall
status using a set of predictor values, including enrollment, duration, phase,
intervention type and source class through a logistic regression model
}
