% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summarize_crc_table}
\alias{summarize_crc_table}
\title{Summary table of clinical study types, intervention types, and phases}
\usage{
summarize_crc_table(dates = c("2000-01-01", "2100-01-01"))
}
\arguments{
\item{dates}{A list of two date values (lower and upper, respectively)}
}
\value{
A summary table showing the clinical trials distribution
}
\description{
Filters mCRC clinical trials within a given range of trial start date
and creates a summary table showing the distribution of study types stratified
by intervention types and phases
}
