% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{title_kw_search}
\alias{title_kw_search}
\title{Function that filters the studies data based on brief title keywords queries}
\usage{
title_kw_search(studies, kw)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{kw}{A string. The keyword string entered by the users}
}
\value{
A tibble with filtered entries
}
\description{
Function that filters the studies data based on brief title keywords queries
}
