% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_kwds}
\alias{query_kwds}
\title{Function that queries a data table using a keyword/a list of keyword strings}
\usage{
query_kwds(tbl, kwds, column, ignore_case = TRUE, match_all = FALSE)
}
\arguments{
\item{tbl}{A tibble. The data table to be filtered}

\item{kwds}{A string / list of strings. The query keywords}

\item{column}{A string. The data column to be filtered upon.}

\item{ignore_case}{A Boolean, default TRUE. Specify whether the query is case-insensitive}

\item{match_all}{A Boolean, default FALSE. Specify whether the query
requires the column value to match all the kwds or just one of them.}
}
\value{
A tibble with filtered entries
}
\description{
Function that queries a data table using a keyword/a list of keyword strings
}
